+++ android/app/src/test/kotlin/com/vscodroid/setup/LaunchRepairWiringTest.kt
@@ -78,6 +78,7 @@ class LaunchRepairWiringTest {
      */
     private val repairs = listOf(
         "setupToolSymlinks",
+        "setupOpenCode",
         "setupGitCore",
         "setupGitCaBundle",
         "setupRipgrepVscodeSymlink",